14_10_makeworking

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: pgomez-r <pgomez-r@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/05 14:45:48 by pgomez-r          #+#    #+#             */
-/*   Updated: 2022/10/13 14:03:53 by pgomez-r         ###   ########.fr       */
+/*   Updated: 2022/10/14 09:33:40 by pgomez-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	char	*dest_cpy;
-	char	*src_cpy;
-	int		i;
+	char		*dest_cpy;
+	char		*src_cpy;
+	size_t		i;
 
 	if (dest == NULL)
 		return (NULL);
